+++ cider-mode.el
@@ -209,6 +209,8 @@ Returns to the buffer in which the command was invoked."
     (define-key map (kbd "C-c ,")   #'cider-test-run-tests)
     (define-key map (kbd "C-c C-,") #'cider-test-rerun-tests)
     (define-key map (kbd "C-c M-,") #'cider-test-run-test)
+    (define-key map (kbd "C-c M-<") #'cider-test-run-loaded-tests)
+    (define-key map (kbd "C-c C-<") #'cider-test-run-project-tests)
     (define-key map (kbd "C-c C-t") #'cider-test-show-report)
     (define-key map (kbd "C-c M-s") #'cider-selector)
     (define-key map (kbd "C-c M-r") #'cider-rotate-default-connection)
@@ -244,7 +246,9 @@ Returns to the buffer in which the command was invoked."
          ["Go back" cider-pop-back])
         ("Test"
          ["Run test" cider-test-run-test]
-         ["Run all tests" cider-test-run-tests]
+         ["Run namespace tests" cider-test-run-tests]
+         ["Run all loaded tests" cider-test-run-loaded-tests]
+         ["Run all project tests" cider-test-run-project-tests]
          ["Rerun failed/erring tests" cider-test-rerun-tests]
          ["Show test report" cider-test-show-report])
         "--"